+++ src/app/cart/components/process-order/process-order.component.ts
@@ -1,5 +1,8 @@
-import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+import { Component, OnInit, OnDestroy, ChangeDetectionStrategy } from '@angular/core';
 import { FormArray, FormBuilder, FormGroup, FormControl, Validators, AbstractControl } from '@angular/forms';
+
+import { debounceTime } from 'rxjs/operators/debounceTime';
+import { Subscription } from 'rxjs/Subscription';
 
 @Component({
   selector: 'app-process-order',
@@ -7,38 +10,104 @@ import { FormArray, FormBuilder, FormGroup, FormControl, Validators, AbstractCon
   styleUrls: ['./process-order.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ProcessOrderComponent implements OnInit {
+export class ProcessOrderComponent implements OnInit, OnDestroy {
   orderForm: FormGroup;
   message: string;
+  emailMessage: string;
+  fullNameMessage: string;
+  phoneMessage: string;
+
+  private sub: Subscription;
+  private validationMessages = {
+    emailMessage: {
+      email: 'Please enter a valid email address.',
+      required: 'Please enter your email address.',
+    },
+    fullNameMessage: {
+      minlength: 'Please enter a valid full name.',
+      required: 'Please enter your full name.',
+    },
+    phoneMessage: {
+      pattern: 'Please enter a valid phone number.',
+      required: 'Please enter your phone number.',
+    }
+  };
 
   constructor(private fb: FormBuilder) { }
 
   ngOnInit() {
-    this.bulder();
+    this.createForm();
+    this.watchValueChanges();
   }
 
-  private bulder() {
+  ngOnDestroy() {
+    this.sub.unsubscribe();
+  }
+
+  private createForm() {
     this.orderForm = this.fb.group({
-      name: new FormControl('', { validators: [Validators.required, Validators.minLength(5)], updateOn: 'blur' }),
+      fullName: new FormControl('',
+        { validators: [Validators.required, Validators.minLength(2), Validators.maxLength(30)],
+          updateOn: 'blur' }),
       email: new FormControl('', { validators: [Validators.required, Validators.email], updateOn: 'blur' }),
+      address: this.fb.group({
+          country: [''],
+          city: [''],
+          zip: [null, [Validators.pattern(/^\d{5}(?:[-\s]\d{4})?$/)]],
+          location: [''],
+      }),
       phones: this.fb.array([this.buildPhone()])
     });
   }
 
   private buildPhone(): FormGroup {
     return this.fb.group({
-      phone: new FormControl('', {validators: [Validators.required, Validators.minLength(7)], updateOn: 'blur'}),
+      phone: new FormControl('', {validators: [Validators.required, Validators.pattern(/^\d{3}-\d{3}-\d{4}$/)], updateOn: 'blur'}),
     });
+  }
+
+  private watchValueChanges(): void {
+    const email = this.orderForm.get('email');
+    const fullName = this.orderForm.get('fullName');
+    const phones = this.orderForm.get('phones');
+
+    this.sub = email.valueChanges
+      .pipe(
+        debounceTime(1000)
+      )
+      .subscribe(() => this.setMessage(email, 'emailMessage'));
+
+    const subFullName = fullName.valueChanges
+      .pipe(
+        debounceTime(1000)
+      )
+      .subscribe(() => this.setMessage(fullName, 'fullNameMessage'));
+
+    let subPhones;
+    phones['controls'].map(phone => subPhones = phone.valueChanges
+        .pipe(
+          debounceTime(1000)
+        )
+        .subscribe(() => this.setMessage(phone, 'phoneMessage'))
+    );
+    this.sub.add(subFullName);
+    this.sub.add(subPhones);
+
+  }
+
+  private setMessage(c: AbstractControl, field: string): void {
+    this[field] = '';
+    if ((c.touched || c.dirty) && (c.errors || c['controls'] && c['controls'].phone.errors)) {
+      const errors = c.errors || c['controls'].phone.errors;
+      this[field] = Object
+        .keys(errors)
+        .map(key => this.validationMessages[field][key])
+        .join(' ');
+    }
   }
 
   get phones(): FormArray {
     return <FormArray>this.orderForm.get('phones');
-  }
-
-  getEmailErrorMessage() {
-    return this.orderForm.controls.email.hasError('required') ? 'You must enter a value' :
-        this.orderForm.controls.email.hasError('email') ? 'Not a valid email' :
-            '';
   }
 
   order(): void {
@@ -50,7 +119,9 @@ export class ProcessOrderComponent implements OnInit {
   }
 
   deletePhone(index): void {
-    this.phones.removeAt(index);
+    if (index !== 0) {
+      this.phones.removeAt(index);
+    }
   }
 
 }